% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_optimal_chunk_size}
\alias{calculate_optimal_chunk_size}
\title{Calculate Optimal Chunk Size}
\usage{
calculate_optimal_chunk_size(
  data_size_mb,
  total_rows,
  max_ram_mb,
  target_fraction = 0.1
)
}
\arguments{
\item{data_size_mb}{Size of data in MB}

\item{total_rows}{Total number of rows}

\item{max_ram_mb}{Maximum RAM in MB}

\item{target_fraction}{Fraction of max RAM to use per chunk (default 0.1)}
}
\value{
Optimal chunk size (number of rows)
}
\description{
Calculates optimal chunk size based on data size and RAM limits.
}
\keyword{internal}
