% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_parallel.R
\name{detect_parallel_backend}
\alias{detect_parallel_backend}
\title{Detect the best parallelization backend for the current environment}
\usage{
detect_parallel_backend()
}
\value{
A list containing backend information:
\item{backend}{Character string identifying the backend. One of:
"mclapply", "parLapply", "doParallel", "doMC", "future",
"foreach", or "sequential"}
\item{os_type}{Operating system type (unix or windows)}
\item{available_cores}{Number of available CPU cores}
\item{packages}{List of available parallel packages}
}
\description{
Detect the best parallelization backend for the current environment
}
\seealso{
\code{\link[=setup_parallel]{setup_parallel()}}, \code{\link[=smart_parallel_apply]{smart_parallel_apply()}}

Other parallel:
\code{\link[=print_parallel_info]{print_parallel_info()}},
\code{\link[=setup_parallel]{setup_parallel()}},
\code{\link[=smart_parallel_apply]{smart_parallel_apply()}},
\code{\link[=stop_parallel]{stop_parallel()}}
}
\concept{parallel}
\keyword{internal}
