% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flir_rules.R
\name{get_flir_rules}
\alias{get_flir_rules}
\title{Get flir rules from fmisc package}
\usage{
get_flir_rules()
}
\value{
Character string with the path to the rules directory.
}
\description{
Returns the path to the flir rules directory in the fmisc package.
These rules can be used in flir/config.yml by adding fmisc to the
\code{from-package} field.
}
\details{
To use these rules in your project, add the following to your
\code{flir/config.yml} file:

\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{from-package:
  - fmisc
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
# Get the rules directory
get_flir_rules()

# Use in flir
library(flir)
setup_flir()
# Then edit flir/config.yml to add fmisc to from-package
}

}
