% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_ram_usage_cpp}
\alias{get_ram_usage_cpp}
\title{Get RAM Usage (Fast C++ Implementation)}
\usage{
get_ram_usage_cpp()
}
\value{
Numeric value representing RAM usage in MB
}
\description{
Fast implementation to get current RAM usage in MB.
}
