% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/di_container.R
\name{print.fmisc_di_container}
\alias{print.fmisc_di_container}
\title{Print an fmisc DI container}
\usage{
\method{print}{fmisc_di_container}(x, ...)
}
\arguments{
\item{x}{A container returned by \code{\link[=di_container]{di_container()}}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an fmisc DI container
}
\seealso{
Other di: 
\code{\link{di_container}()},
\code{\link{fmisc_di}},
\code{\link{with_di_overrides}()}
}
\concept{di}
