% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ram_threshold_exceeded}
\alias{ram_threshold_exceeded}
\title{Monitor RAM During Processing}
\usage{
ram_threshold_exceeded(max_ram_mb)
}
\arguments{
\item{max_ram_mb}{Maximum RAM threshold in MB}
}
\value{
Boolean indicating if RAM threshold is exceeded
}
\description{
Monitors RAM usage and returns true if threshold is exceeded.
}
