% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{split_vector_chunks}
\alias{split_vector_chunks}
\title{Fast Chunk Splitting}
\usage{
split_vector_chunks(x, chunk_size)
}
\arguments{
\item{x}{Numeric vector to chunk}

\item{chunk_size}{Size of each chunk}
}
\value{
List of numeric vectors (chunks)
}
\description{
Efficiently splits a numeric vector into chunks.
}
\keyword{internal}
