% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorators.R
\name{undecorate}
\alias{undecorate}
\title{Unwrap a decorated function by one layer}
\usage{
undecorate(f)
}
\arguments{
\item{f}{A function possibly wrapped by one of the \verb{with_*()} decorators.}
}
\value{
The inner function \code{f} wrapped, or \code{f} unchanged if it is not
a decorated function.
}
\description{
Unwrap a decorated function by one layer
}
\seealso{
Other decorators:
\code{\link[=cache_clear]{cache_clear()}},
\code{\link[=decorate]{decorate()}},
\code{\link{fmisc_decorators}},
\code{\link[=is_decorated]{is_decorated()}},
\code{\link[=with_cache]{with_cache()}},
\code{\link[=with_logging]{with_logging()}},
\code{\link[=with_rate_limit]{with_rate_limit()}},
\code{\link[=with_retry]{with_retry()}},
\code{\link[=with_timing]{with_timing()}}
}
\concept{decorators}
