% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_parallel.R
\name{stop_parallel}
\alias{stop_parallel}
\title{Stop parallel backend and clean up resources}
\usage{
stop_parallel(setup)
}
\arguments{
\item{setup}{The setup object returned by \code{\link{setup_parallel}}}
}
\description{
Stop parallel backend and clean up resources
}
\examples{
\donttest{
setup <- setup_parallel(n_cores = 2)
# ... do parallel work ...
stop_parallel(setup)
}
}
\seealso{
\code{\link{setup_parallel}}, \code{\link{smart_parallel_apply}}

Other parallel:
\code{\link{detect_parallel_backend}()},
\code{\link{print_parallel_info}()},
\code{\link{setup_parallel}()},
\code{\link{smart_parallel_apply}()}
}
\concept{parallel}
